+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - (Breaking change; cli) The verbs `rm` and `remove` have been deleted from all commands (e.g. `forklift pallet rm`), because the two-character verb `rm` doesn't line up nicely with the three-character verb `add`; `del` or `delete` should be used instead (e.g. `forklift pallet del`).
+- (Breaking change; spec) Now the bundle manifest file's `downloads` section lists different download types (e.g. HTTP file download vs. OCI image download) separately, instead of merging them all into one list.
+
+### Removed
+
+- (cli) Now empty lists are omitted from bundle manifest files.
 
 ### Fixed